requirements.txt
alembic==1.6.5
aniso8601==9.0.1
attrs==21.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
Flask==1.1.4
Flask-Env==2.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
flask-restx==0.4.0
Flask-Sitemap==0.3.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
jwt==1.2.0
Mako==1.1.4
MarkupSafe==2.0.1
pycparser==2.20
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.17
urllib3==1.26.5
Werkzeug==0.16.0
